Move api/server/router to daemon/server/router
