+++ api/server/server.go
@@ -8,9 +8,9 @@ import (
 	"github.com/docker/docker/api/server/httpstatus"
 	"github.com/docker/docker/api/server/httputils"
 	"github.com/docker/docker/api/server/middleware"
-	"github.com/docker/docker/api/server/router"
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/versions"
+	"github.com/docker/docker/daemon/server/router"
 	"github.com/docker/docker/dockerversion"
 	"github.com/docker/docker/internal/otelutil"
 	"github.com/gorilla/mux"
