+++ daemon/command/daemon.go
@@ -22,7 +22,6 @@ import (
 	apiserver "github.com/docker/docker/api/server"
 	buildbackend "github.com/docker/docker/api/server/backend/build"
 	"github.com/docker/docker/api/server/middleware"
-	"github.com/docker/docker/api/server/router"
 	buildkit "github.com/docker/docker/builder/builder-next"
 	"github.com/docker/docker/builder/builder-next/exporter"
 	"github.com/docker/docker/builder/dockerfile"
@@ -32,6 +31,7 @@ import (
 	"github.com/docker/docker/daemon/command/trap"
 	"github.com/docker/docker/daemon/config"
 	"github.com/docker/docker/daemon/listeners"
+	"github.com/docker/docker/daemon/server/router"
 	"github.com/docker/docker/daemon/server/router/build"
 	checkpointrouter "github.com/docker/docker/daemon/server/router/checkpoint"
 	"github.com/docker/docker/daemon/server/router/container"
