+++ daemon/server/router/container/container.go
@@ -2,7 +2,7 @@ package container
 
 import (
 	"github.com/docker/docker/api/server/httputils"
-	"github.com/docker/docker/api/server/router"
+	"github.com/docker/docker/daemon/server/router"
 )
 
 // containerRouter is a router to talk with the container controller
