+++ daemon/server/router/grpc/grpc.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/containerd/containerd/v2/defaults"
 	"github.com/containerd/log"
-	"github.com/docker/docker/api/server/router"
+	"github.com/docker/docker/daemon/server/router"
 	"github.com/docker/docker/internal/otelutil"
 	"github.com/moby/buildkit/util/grpcerrors"
 	"github.com/moby/buildkit/util/stack"
